requirements.txt
asgiref==3.8.1
Django==4.2.5
djangorestframework==3.14.0
PyJWT==2.9.0
pytz==2024.1
sqlparse==0.5.1
typing_extensions==4.12.2
